Prune template content before HTML export
